refactor(transit-display): split UI helpers into transit-display-ui.ts
Move the UI presentation API out of build-transit-display-data.ts so
the build module is purely about producing board data, and the new
transit-display-ui.ts owns presentation concerns (sort order, row
caps, empty-state resolution).

Moved to transit-display-ui.ts:
- sortTransitDisplayDataWithMetaData (UI ordering of raw displays)
- resolveTransitDisplayState + TransitDisplayStopsState + TransitDisplayStatus
- transitDisplayMaxEntriesFor + MAX_ENTRIES_BY_INFO_LEVEL
- transitDisplayMaxEntriesPerRouteFor + MAX_ENTRIES_PER_ROUTE_BY_INFO_LEVEL

build-transit-display-data.ts now exports CATEGORIES and DIRECTIONS so
the UI sort can keep its axes in sync with the builder's iteration
order. InfoLevel import is dropped (no longer used in this file).

Callers (TransitDisplaysContainer / TransitDisplays / TransitDisplays2)
follow the new import paths; behaviour is unchanged.

Tests split: the three describe blocks for the moved helpers now live
in transit-display-ui.test.ts, alongside fresh tests for
resolveTransitDisplayState.

Refs: #296
